chore: fix "task" e2e test's `AfterSuite` (#5783)
Our "task" e2e test has being failing because we aren't able to empty the bucket. It is because we keep deploying tasks under the same group name; the latter-deployed tasks are in default cluster, which effectively removes the env manager role's permission to manage the bucket.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.